pyproject.toml
[tool.poetry]
name = "chatroom-websocket"
version = "0.1.0"
description = ""
authors = ["WuangNyH <kayzz0974@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.1"
mysql-connector-python = "^8.2.0"
flask-socketio = "^5.3.6"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
bcrypt = "^4.1.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
